+++ includes/Server.hpp
@@ -31,7 +31,7 @@ class Server
 		std::string							_error_pages;
 		std::string							_index;
 		bool								_autoindex;
-		int									_client_body_size;
+		long int							_client_body_size;
 		std::map<std::string, std::string>	_cgi;
 		std::vector<Location>				_location;
 
@@ -70,7 +70,7 @@ class Server
 		std::string					const &getError() const;
 		std::string					const &getIndex() const;
 		bool						const &getAutoindex() const;
-		int							const &getClientBodySize() const;
+		long int					const &getClientBodySize() const;
 		std::vector<Location>		const &getLocation() const;
 		std::map<std::string, std::string>	const &getCgi() const;
